dev-requirements.txt
pipdeptree
ipython
django-debug-toolbar
factory_boy
